+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - The debugging expresson mini-language.
 - A warning is printed when `R11` is used as the second argument of `NOT`.
 - A warning is printed when `CALL` and `RETURN` are used with atypical registers.
+- `help` debugging command takes one or more arguments and prints more detailed help messages for specific commands.
 
 ### Changed
 - Argument to skip command in debugger is now interpreted as number of lines to skip instead of destination to skip to.